Refactor components directory structure and update storybook config
