+++ src/components/complex/codeable-concept/codeable-concept.data.ts
@@ -1,7 +1,7 @@
-import {BaseData}   from '../../../internal/base/BaseData'
-import {CodingData} from '../coding/coding.data'
+import {BaseElementData} from '../../../internal/base/base-element.data'
+import {CodingData}      from '../coding/coding.data'
 
-export type CodeableConceptData = BaseData & {
+export type CodeableConceptData = BaseElementData & {
   coding: CodingData[]
   text?: string
 }
