+++ src/components/complex/identifier/identifier.data.ts
@@ -1,11 +1,11 @@
-import {BaseData}  from '../../../internal/base/BaseData'
-import {Code, URI} from '../../primitive/primitive.data'
+import {BaseElementData} from '../../../internal/base/base-element.data'
+import {Code, URI}       from '../../primitive/primitive.data'
 
 import {ReferenceData}       from '../../special/reference/reference.data'
 import {CodeableConceptData} from '../codeable-concept/codeable-concept.data'
 import {PeriodData}          from '../period/period.data'
 
-export type IdentifierData = BaseData & {
+export type IdentifierData = BaseElementData & {
   use?: Code
   type?: CodeableConceptData
   system?: URI
