+++ src/components/complex/period/period.data.ts
@@ -1,7 +1,7 @@
-import {BaseData} from '../../../internal/base/BaseData'
-import {DateTime} from '../../primitive/primitive.data'
+import {BaseElementData} from '../../../internal/base/base-element.data'
+import {DateTime}        from '../../primitive/primitive.data'
 
-export type PeriodData = BaseData & {
+export type PeriodData = BaseElementData & {
   start?: DateTime
   end?: DateTime
 }
