+++ src/components/complex/period/period.ts
@@ -1,11 +1,11 @@
 import {html, TemplateResult} from 'lit'
 import {customElement}        from 'lit/decorators.js'
 import {PrimitiveType}        from '../../../components/primitive/type-converters'
-import {ConsumerBaseElement}  from '../../../internal/base/ConsumerBaseElement'
+import {BaseElementConsumer}  from '../../../internal/base/base-element-consumer'
 import {PeriodData}           from './period.data'
 
 @customElement('fhir-period')
-export class Period extends ConsumerBaseElement<PeriodData> {
+export class Period extends BaseElementConsumer<PeriodData> {
 
   constructor() {super('Period')}
