--- src/components/resources/backbone.data.ts
@@ -1,6 +1,0 @@
-import {BaseData}  from '../../internal/base/BaseData'
-import {Extension} from './index'
-
-export type BackboneElementData = BaseData & {
-  modifierExtension: Extension[]
-}
