+++ src/components/special/reference/reference.data.ts
@@ -1,8 +1,8 @@
-import {BaseData}       from '../../../internal/base/BaseData'
-import {IdentifierData} from '../../complex/identifier/identifier.data'
-import {URI}            from '../../primitive/primitive.data'
+import {BaseElementData} from '../../../internal/base/base-element.data'
+import {IdentifierData}  from '../../complex/identifier/identifier.data'
+import {URI}             from '../../primitive/primitive.data'
 
-export type ReferenceData = BaseData & {
+export type ReferenceData = BaseElementData & {
   reference?: string
   type?: URI
   identifier?: IdentifierData
