--- src/internal/base/BaseData.ts
@@ -1,4 +1,0 @@
-export type BaseData = {
-  id?: string | null,
-  extension?: [],
-}
