--- src/internal/base/BaseElementMode.ts
@@ -1,7 +1,0 @@
-export enum BaseElementMode {
-  display = 'display',
-  structure = 'structure',
-  structure_trace = 'structure_trace',
-  combined = 'combined',
-  narrative = 'narrative'
-}
