+++ src/internal/contexts/context.ts
@@ -1,8 +1,8 @@
-import {createContext} from '@lit/context'
-import {ResourceData}  from '../components/resources'
+import {createContext}   from '@lit/context'
+import {BaseElementMode} from '../base/base-element.data'
 
+import {ResourceData} from '../resource/domain-resource.data'
 
-import {BaseElementMode} from '../internal/base/BaseElementMode'
 
 export const containedDataContext = createContext<ResourceData[]>(Symbol('fhir-contained-data'))
 export type DisplayConfig = { mode: BaseElementMode, showerror: boolean, verbose: boolean, open: boolean }
