+++ src/internal/index.ts
@@ -1,7 +1,9 @@
-import './base/BaseData'
-import './base/BaseElement'
-import './base/BaseElementMode'
-import './base/ConsumerBaseElement'
-import './base/ProviderBaseElement'
+import './base/base-element.data'
+import './base/base-element'
+import './base/base-element-consumer'
+import './base/base-element-provider'
 
 import './resource/renderResourceComponent'
+import './resource/domain-resource'
+import './resource/backbone.data'
+import './resource/domain-resource.data'
