+++ src/internal/resource/backbone.data.ts
@@ -1,0 +1,5 @@
+import {BaseElementData, Extension} from '../base/base-element.data'
+
+export type BackboneElementData = BaseElementData & {
+  modifierExtension: Extension[]
+}
