+++ src/shell/layout/Wrapper.ts
@@ -2,7 +2,7 @@ import {consume}                                                   from '@lit/co
 import {css, html, nothing}                                        from 'lit'
 import {customElement, property}                                   from 'lit/decorators.js'
 import {classMap}                                                  from 'lit/directives/class-map.js'
-import {defaultDisplayConfig, DisplayConfig, displayConfigContext} from '../../contexts/context'
+import {defaultDisplayConfig, DisplayConfig, displayConfigContext} from '../../internal/contexts/context'
 import {FhirElement}                                               from '../FhirElement'
